refactorised postcontroller and postservice and created ownershipvalidator service

+++ backend/src/main/java/ro/neforii/service/PostService.java
@@ -1,11 +1,13 @@
 package ro.neforii.service;
 
+import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.stereotype.Service;
 import ro.neforii.dto.post.PostRequestDto;
 import ro.neforii.dto.post.PostResponseDto;
 import ro.neforii.dto.post.PostUpdateRequestDto;
 import ro.neforii.dto.post.PostVoteResponseDto;
 import ro.neforii.dto.vote.VoteRequestDto;
+import ro.neforii.exception.ForbiddenActionException;
 import ro.neforii.exception.PostNotFoundException;
 import ro.neforii.mapper.PostMapper;
 import ro.neforii.model.Post;
@@ -13,6 +15,7 @@ import ro.neforii.model.User;
 import ro.neforii.model.VoteType;
 import ro.neforii.repository.PostRepository;
 import ro.neforii.repository.VoteRepository;
+import ro.neforii.service.security.OwnershipValidator;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -22,24 +25,27 @@ import java.util.UUID;
 public class PostService {
 
     private final PostRepository postRepository;
-    private final VoteRepository voteRepository;
     private final VoteService voteService;
     private final UserService userService;
     private final PostMapper postMapper;
+    private final OwnershipValidator ownershipValidator;
 
 
-    public PostService(PostRepository postRepository, VoteRepository voteRepository, PostMapper postMapper, UserService userService, VoteService voteService) {
+    public PostService(PostRepository postRepository, PostMapper postMapper, UserService userService, VoteService voteService,
+                       OwnershipValidator ownershipValidator) {
         this.postRepository = postRepository;
-        this.voteRepository = voteRepository;
         this.postMapper = postMapper;
         this.userService = userService;
         this.voteService = voteService;
+        this.ownershipValidator = ownershipValidator;
     }
 
 
-    public PostResponseDto createPost(PostRequestDto postRequestDto) {
+    public PostResponseDto createPost(PostRequestDto postRequestDto, UUID currentUserId) {
         User user = userService.getUserByUsername(postRequestDto.author());
         Post post = postMapper.postRequestDtoToPost(postRequestDto, user);
+
+//        ownershipValidator.assertPostOwner(currentUserId, post);
 
         Post savedPost = postRepository.save(post);
 
@@ -61,15 +67,11 @@ public class PostService {
         return postMapper.postToPostResponseDto(post, currentUserId);
     }
 
-    public List<Post> findAll() {
-        return postRepository.findAll();
-    }
-
     public PostResponseDto updatePost(UUID id, PostUpdateRequestDto postUpdateRequestDto, UUID currentUserId) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Post with ID " + id + " not found."));
 
-        // TODO cand avem autentificare putem verifica daca userul curent este cel care a creat postarea
+//        ownershipValidator.assertPostOwner(currentUserId, post);
 
         if (postUpdateRequestDto.title() != null) {
             post.setTitle(postUpdateRequestDto.title());
@@ -87,28 +89,25 @@ public class PostService {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Post with ID " + id + " not found."));
 
-        // TODO cand avem autentificare putem verifica daca userul curent este cel care a creat postarea
-//        if (!post.getUser().getId().equals(currentUserId)) {
-//            throw new PostNotFoundException("You are not authorized to delete this post.");
-//        }
+        if (!post.getUser().getId().equals(currentUserId)) {
+            throw new ForbiddenActionException("The current user is not the author of the post.");
+        }
 
         postRepository.delete(post);
     }
 
     public PostVoteResponseDto votePost(UUID postId, UUID currentUserId, VoteRequestDto voteRequestDto) {
-//      TODO userId la autentificare si award la un anumit numar de voturi
         VoteType voteType = VoteType.fromString(voteRequestDto.voteType());
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new PostNotFoundException("Post with ID " + postId + " not found."));
 
-        voteService.deleteByUserIdAndPostId(currentUserId, postId);
+//        ownershipValidator.assertPostOwner(currentUserId, post);
 
-        if (voteType != VoteType.NONE) {
-            voteService.createVote(currentUserId, postId, null, voteType == VoteType.UP);
-        }
+        voteService.createOrUpdateVoteForPost(currentUserId, postId, voteType);
 
         Post updatedPost = postRepository.findById(postId)
                 .orElseThrow(() -> new PostNotFoundException("Post with ID " + postId + " not found."));
 
         return postMapper.postToPostVoteResponseDto(updatedPost, currentUserId);
     }
-
 }
